vfs: Move forward declarations to separate file
